perf: render weather/cloud shaders at mediump precision
Drop highp float to mediump in the weather sky, card-water, rain-on-glass,
and sun-flare fragments — mediump is sufficient for their colour work and
is cheaper on tile-based GPUs.

+++ shaders/weather/rain_on_glass.frag
@@ -42,7 +42,7 @@
 //   sampler 0: uTex          the content being filtered — SET BY THE ENGINE
 #include <flutter/runtime_effect.glsl>
 
-precision highp float;
+precision mediump float;
 
 uniform vec2 uSize;
 uniform vec2 uResolution;
